Fix module example naming
In the modules affected, the name of the module in the example is not
correct. This is probably due to renaming in the past, but can be
confusing when using the modules.

This commit fixes the naming of the modules in the example section of
the modules to reflect the current name of the module.

Signed-off-by: Simon Krenger <simon.krenger@swisscom.com>