Changing the license from BSD-2-Clause to BSD-2-Clause or GPLv3
The change is specifically to add an additional license under which the
code will be licensed (going forward it will be dual licensed),
specifically to add GPLv3 in addition to it's existing license.  This
change is being done to better align the code with upstream Ansible.

Cherrypicked the master change into dev branch. Added clause in
newly added modules too.
Signed-off-by: John 'Warthog9' Hawley <jhawley@vmware.com>
Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>
Signed-off-by: Kommireddy Akhilesh<akhileshkommireddy2412@gmail.com>