Enhancement: Support sub-resource list specifications
Until now, user had to prepent a prefix to specify any subresource
spec in the playbook. This limited the user's scope to specify only
1 subresource. However, it is possible that 2 or more subresources
can be associated with a single resource. This change
modifies the behavior in a way that the user can specify multiple
subresources under its respective section, which is clear from the
Ansible Module's doc.

Additionally, there are some speed enhancements:
1. We no longer create an Ansible Module for each subresource
2. We no longer test and create Ansible Module to infer which
   resources are specified by the user at runtime. This is possible
   with the help of using spec_identifiers

Lastly, using spec_identifiers improves the code readability as well
as user experience.

Issue: #123

Signed-off-by: Gautam Verma <vermag@vmware.com>
Signed-off-by: Gautam Verma <gautam94verma@gmail.com>