Node_id removed and revision issue
Before FN-TN unification, fabric node id was a mandatory field
but now it is not. Today node_id is same as that of tn id.
Thus there is no point in having that. Thus removing node_id.
node_id given is same as that of TN id.
The module was not getting updated due as revision was required
in node_deployment_request field also. This CL solves it.

Testing done:
Created a Host TN successfully.
Updated its description. 4 times successfuly.
Deleted the TN successfully.

Signed-off-by: Kommireddy Akhilesh<akhileshkommireddy2412@gmail.com>