Move Policy modules to Ansible Base
Until now, all the work related to Policy was being done in the
directory policy-ansible-for-nsxt. This patch merges/moves the
Policy code into the base directory so that only one Ansible
repository is formed. This makes it easier to use and allows us
to move to Ansible Collections and Galaxy

Signed-off-by: Gautam Verma <gautam94verma@gmail.com>