Upgrade pre check break
Upgrade pre check was failing to retrieve data from an API,
as API was responding in csv format. This issue is not
happening in python 2 but occurring in Python3 due to
type changes in python3 response.
Given that there were previous suggestions about not showing
csv format on the ansible output as it is confusing. Showing
the upgrade results from upgrade/pre-upgrade-checks/failures
rather than upgrade/pre-upgrade-checks?format=csv
This works from both versions of python 2 and 3.

Signed-off-by: Kommireddy Akhilesh<akhileshkommireddy2412@gmail.com>