Remove empty params for req body
We should remove empty params from request body while making PATCH
call if the same attrs are not set already on the resource.
Otherwise, we assume that there is a PATCH operation
being done on it while in reality, there is None.

Eg. Policy API returns a response in which None attrs are not set.
When we create the user request body, we set attrs to None. Although
both mean semantically the same, the syntax difference makes us
assume that a PATCH call should be done

Issue: #317