Verify resource was updated after PATCH
In some cases, it is not possible to infer if a PATCH call should
be made on the NSX-t resource. This is because Ansible substitues
missing keys with None. We always send the PATCH call in such cases.
If the server updates the resource, it's revision is updated and
we check it to infer if it was updated actually.

Issue: #2658008

Signed-off-by: Gautam Verma <vermag@vmware.com>
Signed-off-by: Gautam Verma <gautam94verma@gmail.com>