Infer LocaleService ID automatically if precreated
LocaleService is automatically created by NSX when the user tries
to do some operation on tier0/1 like set the Edge Cluster. In such cases,
the ID cannot be known by the user as they only have access to UI.
Therefore, if it cannot be specified we use the existing LocaleService
on the gateway and configure that.

If no LocaleService was present but user didn't specify ID or display_name,
then error is thrown to provide either of them

If multiple LocaleServices are present, error is thrown to specify the exact
LocaleService ID.

Issue: #2655611