Fix reading resource params with nested spec
In Policy modules, the resources are nested. Because of this, we read
the resource params from the supplied Ansible Playbook and manually
validate it with the resource spec. However, we were not reading the
resource params correctly if they were nested. This patch fixes that
issue. In addition, we convert the new resource params and existing
resource params into Ordered Data so that we can compare them and infer
if there are any updates with better accuracy

Issue: #417

Signed-off-by: Gautam Verma <vermag@vmware.com>
Signed-off-by: Gautam Verma <gautam94verma@gmail.com>